Implemented visuals, showing player and enemy Armies moving.

+++ AI-FuzzyLogic/Fuzzy.cpp
@@ -7,9 +7,9 @@
 /// <param name="t_x0">Leftmost Value (0)</param>
 /// <param name="t_x1">Rightmost Value (1)</param>
 /// <returns>Value position in Grade</returns>
-double FuzzyLogic::Grade(double t_val, double t_x0, double t_x1)
+float FuzzyLogic::Grade(float t_val, float t_x0, float t_x1)
 {
-	double result = 0;
+	float result = 0;
 
 	if (t_val <= t_x0)
 		result = 0;
@@ -29,9 +29,9 @@ double FuzzyLogic::Grade(double t_val, double t_x0, double t_x1)
 /// <param name="t_x1">Center Value (1)</param>
 /// <param name="t_x2">Rightmost Value (0)</param>
 /// <returns>Value position in Triangle</returns>
-double FuzzyLogic::Triangle(double t_val, double t_x0, double t_x1, double t_x2)
+float FuzzyLogic::Triangle(float t_val, float t_x0, float t_x1, float t_x2)
 {
-	double result = 0;
+	float result = 0;
 
 	if ((t_val <= t_x0) || (t_val >= t_x2))
 		result = 0;
@@ -54,9 +54,9 @@ double FuzzyLogic::Triangle(double t_val, double t_x0, double t_x1, double t_x2)
 /// <param name="t_x2">Center Rightmost Value (1)</param>
 /// <param name="t_x3">Rightmost Value (0)</param>
 /// <returns>Value position in Trapezoid</returns>
-double FuzzyLogic::Trapezoid(double t_val, double t_x0, double t_x1, double t_x2, double t_x3)
+float FuzzyLogic::Trapezoid(float t_val, float t_x0, float t_x1, float t_x2, float t_x3)
 {
-	double result = 0;
+	float result = 0;
 
 	if ((t_val <= t_x0) || (t_val >= t_x3))
 		result = 0;
