+++ AI-FuzzyLogic/Fuzzy.h
@@ -14,14 +14,14 @@ public:
 	FuzzyLogic() {};
 
 	// Membership Functions
-	static double Grade(double t_val, double t_x0, double t_x1);
-	static double Triangle(double t_val, double t_x0, double t_x1, double t_x2);
-	static double Trapezoid(double t_val, double t_x0, double t_x1, double t_x2, double t_x3);
+	static float Grade(float t_val, float t_x0, float t_x1);
+	static float Triangle(float t_val, float t_x0, float t_x1, float t_x2);
+	static float Trapezoid(float t_val, float t_x0, float t_x1, float t_x2, float t_x3);
 
 	// Logical Operators
-	static double AND(double t_a, double t_b) { return std::min(t_a, t_b); };
-	static double OR(double t_a, double t_b) { return std::max(t_a, t_b); };
-	static double NOT(double t_a) { return 1.0 - t_a; };
+	static float AND(float t_a, float t_b) { return std::min(t_a, t_b); };
+	static float OR(float t_a, float t_b) { return std::max(t_a, t_b); };
+	static float NOT(float t_a) { return 1.0f - t_a; };
 };
 
 #endif
